pyproject.toml
[project]
name = "chores_agent"
version = "0.1"
description = "AI agent to divide, assign, and remind household chores among roommates"
authors = [{ name = "Anushka Shukla" }]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^1.0.0"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = { version = "^1.93", extras = [
    "adk",
    "agent-engines",
] }
mysql-connector-python = "^8.4.0"  # assuming you're using MySQL

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
google-cloud-aiplatform = { version = "^1.93", extras = [
    "adk",
    "agent-engines",
    "evaluation",
] }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
